requirements.txt
azure-ai-documentintelligence
azure-identity
python-dotenv
streamlit
PyPDF2
spacy
google-generativeai